Rename Hazel/Core/Core.h to Hazel/Core/Base.h per episode 49. This is to match Hazel-dev.
