+++ Hazel/src/Hazel.h
@@ -2,7 +2,7 @@
 
 // For use by Hazel applications
 
-#include "Hazel/Core/Core.h"
+#include "Hazel/Core/Base.h"
 
 #include "Hazel/Core/Application.h"
 #include "Hazel/Core/Layer.h"
