+++ Hazel/src/Hazel/Core/Base.h
