+++ Hazel/src/Hazel/Core/Layer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Hazel/Core/Core.h"
+#include "Hazel/Core/Base.h"
 #include "Hazel/Core/Timestep.h"
 #include "Hazel/Events/Event.h"
