+++ Hazel/src/Hazel/Core/LayerStack.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Hazel/Core/Core.h"
+#include "Hazel/Core/Base.h"
 #include "Hazel/Core/Layer.h"
 
 #include <vector>
