+++ Hazel/src/Hazel/Core/Window.h
@@ -2,7 +2,7 @@
 
 #include "hzpch.h"
 
-#include "Hazel/Core/Core.h"
+#include "Hazel/Core/Base.h"
 #include "Hazel/Events/Event.h"
 
 namespace Hazel {
