+++ Hazel/src/Hazel/Renderer/Texture.h
@@ -2,7 +2,7 @@
 
 #include <string>
 
-#include "Hazel/Core/Core.h"
+#include "Hazel/Core/Base.h"
 
 namespace Hazel {
